modify the predictions.html page
There was a problem with the expected weekly revenue which was shown to be more than the yearly revenue due to some mistake in the calculation. This was fixed. Another thing that was fixed was the behavior of the page when a user inputs an unreasonable price (defined as price > 1000). Previously, an alert was given "infinite" amount of times. This was fixed to ensure that the alert prompting the user to input reasonable price was done only once and all values reset to 0