Modify alert message
Alerts were combined such that they are produced at once (thus not overwhelming the user. Another thing done was to make submission of entries to happen on click. This is important for consistency as on the previous version, some calculations happened on click while for others it happened on submission